REF: move remaining TimedeltaIndex arith tests
