+++ pandas/tests/indexes/timedeltas/test_arithmetic.py
@@ -1,36 +1,13 @@
-from datetime import timedelta
-
-import numpy as np
 import pytest
 
 from pandas.errors import NullFrequencyError
 
 import pandas as pd
-from pandas import Timedelta, TimedeltaIndex, timedelta_range
+from pandas import TimedeltaIndex
 import pandas.util.testing as tm
 
 
-@pytest.fixture(
-    params=[
-        pd.offsets.Hour(2),
-        timedelta(hours=2),
-        np.timedelta64(2, "h"),
-        Timedelta(hours=2),
-    ],
-    ids=str,
-)
-def delta(request):
-    # Several ways of representing two hours
-    return request.param
-
-
-@pytest.fixture(params=["B", "D"])
-def freq(request):
-    return request.param
-
-
-class TestTimedeltaIndexArithmetic:
-    # Addition and Subtraction Operations
+class TestTimedeltaIndexShift:
 
     # -------------------------------------------------------------
     # TimedeltaIndex.shift is used by __add__/__sub__
@@ -96,61 +73,3 @@ class TestTimedeltaIndexArithmetic:
         tdi = TimedeltaIndex(["1 days 01:00:00", "2 days 01:00:00"], freq=None)
         with pytest.raises(NullFrequencyError):
             tdi.shift(2)
-
-    # -------------------------------------------------------------
-    # Binary operations TimedeltaIndex and timedelta-like
-    # Note: add and sub are tested in tests.test_arithmetic, in-place
-    #  tests are kept here because their behavior is Index-specific
-
-    def test_tdi_iadd_timedeltalike(self, delta):
-        # only test adding/sub offsets as + is now numeric
-        rng = timedelta_range("1 days", "10 days")
-        expected = timedelta_range("1 days 02:00:00", "10 days 02:00:00", freq="D")
-        rng += delta
-        tm.assert_index_equal(rng, expected)
-
-    def test_tdi_isub_timedeltalike(self, delta):
-        # only test adding/sub offsets as - is now numeric
-        rng = timedelta_range("1 days", "10 days")
-        expected = timedelta_range("0 days 22:00:00", "9 days 22:00:00")
-        rng -= delta
-        tm.assert_index_equal(rng, expected)
-
-    # -------------------------------------------------------------
-
-    def test_tdi_ops_attributes(self):
-        rng = timedelta_range("2 days", periods=5, freq="2D", name="x")
-
-        result = rng + 1 * rng.freq
-        exp = timedelta_range("4 days", periods=5, freq="2D", name="x")
-        tm.assert_index_equal(result, exp)
-        assert result.freq == "2D"
-
-        result = rng - 2 * rng.freq
-        exp = timedelta_range("-2 days", periods=5, freq="2D", name="x")
-        tm.assert_index_equal(result, exp)
-        assert result.freq == "2D"
-
-        result = rng * 2
-        exp = timedelta_range("4 days", periods=5, freq="4D", name="x")
-        tm.assert_index_equal(result, exp)
-        assert result.freq == "4D"
-
-        result = rng / 2
-        exp = timedelta_range("1 days", periods=5, freq="D", name="x")
-        tm.assert_index_equal(result, exp)
-        assert result.freq == "D"
-
-        result = -rng
-        exp = timedelta_range("-2 days", periods=5, freq="-2D", name="x")
-        tm.assert_index_equal(result, exp)
-        assert result.freq == "-2D"
-
-        rng = pd.timedelta_range("-2 days", periods=5, freq="D", name="x")
-
-        result = abs(rng)
-        exp = TimedeltaIndex(
-            ["2 days", "1 days", "0 days", "1 days", "2 days"], name="x"
-        )
-        tm.assert_index_equal(result, exp)
-        assert result.freq is None
